FIX: #17 walkaround to allow negative capture window coordinates
The Electron framework seems to have problems when creating windows at negative origin locations. Negative locations will become significant when the user uses multiple monitors and does not configure the most left monitor as the main one. Locating windows on that monitor automatically results in negative coordinates, as the top left corner of the main window will be regarded as (0,0). By first, creating the render window at (0,0) and then doing all the other configurations and afterwards, set the desired location values (which might be negative ones) solves the conflict.